imepoch.c has a new version of the island model of migration.  I think the old version was incorrect.  It did not make the migration process symmetric, so groups would have tended to change in size.  This means that the migration process looks different in forward and backwards time, but the old code took no account of this.  In short, it was botched but now it's fixed.  The new model is explained fully in the comments preceding IMEpoch_hazard.
This change in model requires a change in the interpretation of the
parameter read in from impophist.ini.  Rather than measuring the number of
migrants exchanged by each pair of groups, the migration parameter now
equals the hazard that a gamete will migrate per unit of mutational time.
The average number of migrants into each group should equal unity when this
parameter (m) equals the reciprocal of the average of theta[i].