Fixed a bug in imcoalesce.c:  Extinction events were given wrong dates.
Discovered a bug in event.c.  The problem is that I store branch lengths but
not dates.  Consequently, the dates I use are calculated and contain error.
Then, when the events are put into an EventList and sorted, the sort order is
wrong.  There can be extinction events that occur after the NewEpoch event with
which they are associated, and which consequently involve "migration" out of
empty demes.  The debugging code aborts because the hazard of this event
is zero.