I'm in the middle of coding the case where c is approximately 1 in PrExp::reset.
One problem has materialized: The adolc library does not include an
overloaded function for expm1.  Do I add it myself?  This ought
to be pretty easy, by using the code for this function in R, but
expm1 requires log1p, which required cheb_init and cheb_eval.  I would
need to port all of these to adolc.